[REFACTOR] #267 PostDetail 컴포넌트 폴더 분리
