+++ src/domains/Posts/components/PostDetail/PostDetailManageCard.tsx
