+++ src/domains/Posts/components/PostDetail/PostDetailReportCard.tsx
